requirements/base.txt
Django>=3.2
djangorestframework
pytest
pytest-django
pytest-cov
psycopg2-binary
